cli: make main command easier to call. (#71)
