+++ pkg/cli/cli_test.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	_ "embed"
@@ -52,9 +52,9 @@ func TestBanner(t *testing.T) {
 		require.NoError(t, err)
 
 		if hideBanner {
-			assert.NotContains(t, string(data), banner)
+			assert.NotContains(t, string(data), Banner)
 		} else {
-			assert.Contains(t, string(data), banner)
+			assert.Contains(t, string(data), Banner)
 		}
 	}
